Add files via upload
Shell and improved bubble sort are both optimized. 

Shell sort is using a gap sequence called Three Smooth Number for such optimization. And the time complexity is O(n^2) and the space complexity is O(n).

Improved Bubble sort uses the sequence that is the size keep diving by 1.3. The time and space
complexity is O(n) and O(n).